(#5660) New Type to handle /etc/services entries
This new type "port" handles entries in /etc/services. It uses multiple
key_attributes (name and protocol), so you are able to add e.g.
multiple telnet lines for tcp and udp. Sample usage

    port { 'telnet':
      number      => '23',
      protocol    => 'tcp',
      description => 'Telnet'
    }

Because the type makes use of the title_patterns function this can also
be written as

    port { 'telnet/tcp':
      number      => '23',
      description => 'Telnet'
    }

This type only supports tcp and udp and might not work on OS X.