(#5660) Add integration tests for port provider
Add specs to check the interaction between the port provider and the
catalog and transaction class.

The added specs will currently result in failing tests.

The port provider assumes that the prefetched hash of resources that is
built in transaction.rb uses uniqueness_key as the hashkey. Because that
is not true (yet) puppet has problems whenever it parses /etc/services
and port name is not unique (e.g. telnet/tcp and telnet/udp should be
managed or do appear in the file).